requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Audio processing and transcription
openai-whisper==20231117
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
webrtcvad==2.0.10
noisereduce==3.0.0
azure-cognitiveservices-speech==1.34.0
google-cloud-speech==2.21.0

# Medical NLP and ML
transformers==4.36.2
torch==2.1.2
huggingface-hub==0.19.4
spacy==3.7.2
scispacy==0.5.3
medspacy==0.2.0
datasets==2.15.0
sentence-transformers==2.2.2

# Vector databases
chromadb==0.4.18
pinecone-client==2.2.4
weaviate-client==3.25.3
qdrant-client==1.7.0

# Database and storage
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
redis==5.0.1

# Security and encryption
cryptography==41.0.7
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Utilities
PyYAML==6.0.1
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0

# Form generation and PDF
reportlab==4.0.7
jinja2==3.1.2
weasyprint==60.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
factory-boy==3.3.0

# Development tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Monitoring and logging
structlog==23.2.0
prometheus-client==0.19.0
sentry-sdk==1.38.0

# ICD coding
icd10-cm==0.0.4

# Optional: GPU support (uncomment if using CUDA)
# torch-audio==2.1.2+cu121
# torchvision==0.16.2+cu121